requirements.txt
crewai
crewai-tools
streamlit
python-dotenv
litellm
google-generativeai
